Fix styles of commenting module
