+++ front/app/components/PostShowComponents/Comments/PublicComments.tsx
@@ -34,6 +34,7 @@ import CommentingIdeaDisabled from './CommentingIdeaDisabled';
 
 // utils
 import { isPage } from 'utils/helperUtils';
+import useInitiativesPermissions from 'hooks/useInitiativesPermissions';
 
 const Header = styled(Box)`
   ${isRtl`
@@ -82,6 +83,9 @@ const PublicComments = ({
     ideaId: postType === 'idea' ? postId : undefined,
     sort: sortOrder,
   });
+  const commentingPermissionInitiative = useInitiativesPermissions(
+    'commenting_initiative'
+  );
 
   const commentsList = comments?.pages.flatMap((page) => page.data);
 
@@ -118,14 +122,19 @@ const PublicComments = ({
   const isAdminPage = isPage('admin', pathname);
   const showCommentCount = !isAdminPage && hasComments;
   const showHeader = !isAdminPage || hasComments;
+  const commentingDisabledReason =
+    postType === 'initiative'
+      ? commentingPermissionInitiative?.disabledReason
+      : idea?.data.attributes?.action_descriptor.commenting_idea
+          .disabled_reason;
+  const canComment = !commentingDisabledReason;
 
   return (
     <Box className={className || ''}>
       {showHeader && (
         <Header
           display="flex"
-          flexDirection={isSmallerThanPhone ? 'column' : 'row'}
-          alignItems={isSmallerThanPhone ? 'initial' : 'center'}
+          flexDirection="column"
           justifyContent="space-between"
           mt="16px"
         >
@@ -146,7 +155,7 @@ const PublicComments = ({
             )}
           </>
           {hasComments && (
-            <Box ml="auto">
+            <Box ml="auto" mb="20px">
               <CommentSorting
                 onChange={handleSortOrderChange}
                 selectedCommentSort={sortOrder}
@@ -155,17 +164,17 @@ const PublicComments = ({
           )}
         </Header>
       )}
-
-      <Box my="24px">
-        <ParentCommentForm
-          ideaId={ideaId}
-          initiativeId={initiativeId}
-          postType={postType}
-          postingComment={handleCommentPosting}
-          allowAnonymousParticipation={allowAnonymousParticipation}
-        />
-      </Box>
-
+      {canComment && (
+        <Box mb="24px">
+          <ParentCommentForm
+            ideaId={ideaId}
+            initiativeId={initiativeId}
+            postType={postType}
+            postingComment={handleCommentPosting}
+            allowAnonymousParticipation={allowAnonymousParticipation}
+          />
+        </Box>
+      )}
       <Comments
         ideaId={ideaId}
         initiativeId={initiativeId}
